Use long hex format for colors

+++ packages/stylelint-config-adidas/index.js
@@ -31,7 +31,7 @@ module.exports = {
     'block-opening-brace-space-after': 'always-single-line',
     'block-opening-brace-space-before': 'always',
     'color-hex-case': 'lower',
-    'color-hex-length': 'short',
+    'color-hex-length': 'long',
     'color-named': null,
     'color-no-hex': null,
     'color-no-invalid-hex': true,
